fixes pressure tank merging duplicating gasses + pressure tank buffs (#10969)
## About The Pull Request
fixes a bug with pressure tanks duplicating gasses.

pressure tanks have a volume of 10,000 liters now. 5 canisters worth.

rounded pressure tank pressure limit to 50,000.

halved the construction cost of finishing pressure tank frames.

pressure tank max pressure limit raised to 50
## Why It's Good For The Game
bug fix and made an under utilized atmos component more useful.

## Changelog

:cl:
fix: fixed deconstructing pressure tanks duplicating gasses in merged
tanks.
balance: pressure tanks have a volume of 10,000 liters and rounded the
pressure limit to 50,000.
balance: halved the material cost for finishing pressure tank frames.
/:cl:

## Pre-Merge Checklist

- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [ ] You documented all of your changes.

+++ code/datums/mergers/_merger.dm
@@ -131,7 +131,7 @@
 /datum/merger/proc/check_turf(turf/location, list/found_turfs, asking_from)
 	var/found_something = FALSE
 	// if asking_from is invalid (like if it's 0), we get a random output. that's bad, let's check for falsyness
-	var/us_to_them = asking_from && turn(asking_from, 180)
+	var/us_to_them = asking_from && REVERSE_DIR(asking_from)
 
 	if(found_turfs[location])
 		found_turfs[location][MERGE_TURF_PACKET_DIR] |= us_to_them
@@ -152,8 +152,7 @@
 		found_turfs[location][MERGE_TURF_PACKET_ATOMS] += thing
 		found_something = TRUE
 
-
-	if(merged_typecache[location.type])
+	if(merged_typecache[location.type] && isturf(origin))
 		if(location.mergers && location.mergers[id] != src)
 			var/datum/merger/existing = location.mergers[id]
 			qdel(src)
